♻️ Fixed the ip checker's process
