+++ src-electron/util/network/ip.ts
@@ -1,3 +1,4 @@
+import { z } from 'zod';
 import { Failable } from '../../schema/error';
 import { BytesData } from '../binary/bytesData';
 import { errorMessage } from '../error/construct';
@@ -42,4 +43,41 @@ export async function getGlobalIP(): Promise<Failable<string>> {
   }
 
   return errorMessage.core.failGetGlobalIP();
+}
+
+/** 指定したIPアドレスが有効かどうか */
+export function isValidIP(ip: string): boolean {
+  const parser = z.string().ip();
+  const result = parser.safeParse(ip);
+  return result.success;
+}
+
+/** In Source Testing */
+if (import.meta.vitest) {
+  const { test, expect } = import.meta.vitest;
+
+  test('ip validation', async () => {
+    const ipv4Samples = [
+      '192.168.1.1',
+      '10.0.0.1',
+      '172.16.254.1',
+      '8.8.8.8',
+      '127.0.0.1',
+    ];
+    const ipv6Samples = [
+      '2001:db8:3333:4444:5555:6666:7777:8888',
+      '2001:db8::',
+      '::1',
+      '2001:db8::1234:56',
+      'fe80::394:a9b4:14d:7993',
+      'fe80::1',
+    ];
+
+    for (const ip4 of ipv4Samples) {
+      expect(isValidIP(ip4)).toBe(true);
+    }
+    for (const ip6 of ipv6Samples) {
+      expect(isValidIP(ip6)).toBe(true);
+    }
+  });
 }
